+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Features
 
+- Replace folder_ids with folders in Config
 - Fix ExpertizeMe filename parsing in extract_interlocutor_names
 
 ### Documentation